+++ packages/components/forms/package.json
@@ -8,9 +8,9 @@
   "dependencies": {
     "@babel/runtime": "^7.6.2",
     "@contentful/f36-core": "^4.0.0",
-    "@contentful/f36-helptext": "^4.0.0",
     "@contentful/f36-tokens": "^4.0.0",
-    "@contentful/f36-validation-message": "^4.0.0",
+    "@contentful/f36-icons": "^4.0.0",
+    "@contentful/f36-typography": "^4.0.0",
     "emotion": "^10.0.17"
   },
   "peerDependencies": {
