+++ packages/components/forms/src/help-text/HelpText.mdx
@@ -5,7 +5,7 @@ status: 'stable'
 slug: /components/help-text/
 github: 'https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components/src/components/HelpText'
 storybook: 'https://f36-storybook.contentful.com/?path=/story/components-helptext--default'
-typescript: ./src/HelpText.tsx
+typescript: ./HelpText.tsx
 ---
 
 `HelpText` is a styled copy block with guidance, placed in the context of form components.
