+++ packages/components/forms/src/validation-message/ValidationMessage.mdx
@@ -3,9 +3,9 @@ title: 'ValidationMessage'
 type: 'component'
 status: 'stable'
 slug: /components/validation-message/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/validation-message'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-validationmessage--default'
-typescript: ./src/ValidationMessage.tsx
+github: 'https://github.com/contentful/forma-36/tree/master/packages/components/forms/src/ValidationMessage'
+storybook: 'https://f36-storybook.contentful.com/?path=/story/form--elements-validationmessage--default'
+typescript: ./ValidationMessage.tsx
 ---
 
 `ValidationMessage` provides a consistent styled paragraph for messages in case of validation errors.
