+++ packages/components/forms/stories/ValidationMessage.stories.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import type { Meta, Story } from '@storybook/react/types-6-0';
 
 import { ValidationMessage } from '../src/ValidationMessage';
-import type { ValidationMessageInternalProps } from '../src/ValidationMessage';
+import type { ValidationMessageInternalProps } from '../src/ValidationMessage/ValidationMessage';
 
 export default {
-  title: 'Components/ValidationMessage',
+  title: 'Form Elements/ValidationMessage',
   component: ValidationMessage,
   parameters: {
     propTypes: [ValidationMessage['__docgenInfo']],
