--- packages/components/helptext/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@contentful/f36-helptext",
-  "version": "4.0.0",
-  "description": "Forma 36: HelpText component",
-  "scripts": {
-    "build": "parcel build index.ts"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.6.2",
-    "@contentful/f36-tokens": "^4.0.0",
-    "@contentful/f36-core": "^4.0.0",
-    "@contentful/f36-typography": "^4.0.0",
-    "emotion": "^10.0.17"
-  },
-  "peerDependencies": {
-    "react": ">=16.8"
-  },
-  "license": "MIT",
-  "files": [
-    "dist",
-    "src"
-  ],
-  "main": "dist/main.js",
-  "module": "dist/module.js",
-  "types": "dist/types.d.ts",
-  "source": "src/index.ts",
-  "sideEffects": false,
-  "browserslist": "extends @contentful/browserslist-config",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/contentful/forma-36"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
