--- packages/components/validation-message/README.md
@@ -1,4 +1,0 @@
-# @contentful/f36-validationmessage
-
-This package is part of the pre-release. This means it is unsupported and subject to breaking changes without warning.
-Please use official, supported version of the library [forma-36](https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components), [NPM](https://www.npmjs.com/package/@contentful/forma-36-react-components)
