+++ packages/forma-36-react-components/package.json
@@ -48,8 +48,6 @@
     "@contentful/f36-tokens": "^4.0.0",
     "@contentful/f36-typography": "^4.0.0",
     "@contentful/f36-popover": "^4.0.0",
-    "@contentful/f36-helptext": "^4.0.0",
-    "@contentful/f36-validation-message": "^4.0.0",
     "@contentful/f36-empty-state": "^4.0.0",
     "@contentful/f36-tooltip": "^4.0.0",
     "@contentful/f36-button": "^4.0.0",
